[CLEANUP] Move and rename the basic tests
Move them to the unit/functional tests roots to make
them more discoverable.
