+++ Tests/Functional/ExtensionLoadedTest.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace TTN\Tea\Tests\Functional\Environment;
+namespace TTN\Tea\Tests\Functional;
 
 use PHPUnit\Framework\Attributes\CoversNothing;
 use PHPUnit\Framework\Attributes\Test;
@@ -10,16 +10,17 @@ use TYPO3\CMS\Core\Utility\ExtensionManagementUtility;
 use TYPO3\TestingFramework\Core\Functional\FunctionalTestCase;
 
 #[CoversNothing]
-final class ExtensionTest extends FunctionalTestCase
+final class ExtensionLoadedTest extends FunctionalTestCase
 {
     protected array $testExtensionsToLoad = ['ttn/tea'];
+
     protected bool $initializeDatabase = false;
 
     #[Test]
     public function isLoaded(): void
     {
         self::assertTrue(
-            ExtensionManagementUtility::isLoaded('tea')
+            ExtensionManagementUtility::isLoaded('tea'),
         );
     }
 }
