+++ Tests/Unit/VersionCompatibilityTest.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace TTN\Tea\Tests\Unit\Environment;
+namespace TTN\Tea\Tests\Unit;
 
 use PHPUnit\Framework\Attributes\CoversNothing;
 use PHPUnit\Framework\Attributes\Test;
@@ -10,7 +10,7 @@ use TYPO3\CMS\Core\Information\Typo3Version;
 use TYPO3\TestingFramework\Core\Unit\UnitTestCase;
 
 #[CoversNothing]
-final class ExtensionTest extends UnitTestCase
+final class VersionCompatibilityTest extends UnitTestCase
 {
     #[Test]
     public function currentVersionIsSupported(): void
@@ -19,7 +19,7 @@ final class ExtensionTest extends UnitTestCase
         $currentVersion = (new Typo3Version())->getMajorVersion();
         self::assertContains(
             $currentVersion,
-            $supportedVersions
+            $supportedVersions,
         );
     }
 }
